multicall: deal with early exit conditions
In particular changes to guest privilege level require the multicall
sequence to be aborted, as hypercalls are permitted from kernel mode
only. While likely not very useful in a multicall, also properly handle
the return value in the HYPERVISOR_iret case (which should be the guest
specified value).

This is XSA-213.

Reported-by: Jann Horn <jannh@google.com>
Signed-off-by: Jan Beulich <jbeulich@suse.com>
Reviewed-by: Andrew Cooper <andrew.cooper3@citrix.com>
Acked-by: Julien Grall <julien.grall@arm.com>

Backported to Xen  4.4 for Centos
From: Kevin Stange <kevin@steadfast.net>

Dropped ARM psr_mode_is_user check.  Vulnerability is not present
on ARM, anyway, according to upstream advisory.
Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>


Gbp-Pq: Name multicall-deal-with-early-exit-condition